feat: create function to go upper from current directory
